docsitalia: move metadata related stuff in its own file
